+++ src/main/java/org/dongguk/ownsaemiro/ownsaemiroserver/controller/EventController.java
@@ -1,12 +1,8 @@
 package org.dongguk.ownsaemiro.ownsaemiroserver.controller;
 
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 import org.dongguk.ownsaemiro.ownsaemiroserver.annotation.UserId;
 import org.dongguk.ownsaemiro.ownsaemiroserver.dto.global.ResponseDto;
-import org.dongguk.ownsaemiro.ownsaemiroserver.dto.request.BuyingTicketDto;
-import org.dongguk.ownsaemiro.ownsaemiroserver.dto.request.LikeEventDto;
-import org.dongguk.ownsaemiro.ownsaemiroserver.dto.request.WriteReviewOfEvent;
 import org.dongguk.ownsaemiro.ownsaemiroserver.dto.response.*;
 import org.dongguk.ownsaemiro.ownsaemiroserver.service.EventService;
 import org.springframework.web.bind.annotation.*;
@@ -16,31 +12,10 @@ import org.springframework.web.bind.annotation.*;
 @RequiredArgsConstructor
 public class EventController {
     private final EventService eventService;
-    @PostMapping("/likes")
-    public ResponseDto<?> saveUserLikedEvent(@UserId Long userId, @RequestBody LikeEventDto likeEventDto){
-        LikedEventDto likedEventDto = eventService.userLikeEvent(userId, likeEventDto.id());
 
-        return ResponseDto.created(likedEventDto);
-    }
-
-    @GetMapping("/likes")
-    public ResponseDto<?> showUserLikedEvents(
-            @UserId Long userId,
-            @RequestParam(value = "page", defaultValue = "1") Integer page,
-            @RequestParam(value = "size", defaultValue = "10") Integer size
-    ) {
-        UserLikedEventsDto userLikedEventsDto = eventService.showUserLikedEvents(userId, page-1, size);
-
-        return ResponseDto.ok(userLikedEventsDto);
-    }
-
-    @DeleteMapping("{eventId}/likes")
-    public ResponseDto<?> deleteUserLikedEvent(@UserId Long userId, @PathVariable Long eventId){
-        UnlikedEventDto unlikedEventDto = eventService.userDontLikeEvent(userId, eventId);
-
-        return ResponseDto.ok(unlikedEventDto);
-    }
-
+    /**
+     * 행사 검색하기
+     */
     @GetMapping("/search")
     public ResponseDto<?> searchEvents(
             @RequestParam("name") String name,
@@ -52,6 +27,9 @@ public class EventController {
         return ResponseDto.ok(searchEventsDto);
     }
 
+    /**
+     * 행사 목록 보여주기
+     */
     @GetMapping
     public ResponseDto<?> showEvents(
             @RequestParam("status") String status,
@@ -63,6 +41,10 @@ public class EventController {
 
         return ResponseDto.ok(eventsDto);
     }
+
+    /**
+     * 행사 정보 상세 보기
+     */
 
     @GetMapping("{eventId}/info")
     public ResponseDto<?> showDetailInfoOfEvent(@UserId Long userId, @PathVariable Long eventId){
@@ -78,17 +60,6 @@ public class EventController {
         return ResponseDto.ok(detailDescriptionOfEventDto);
     }
 
-    @PostMapping("{eventId}/review")
-    public ResponseDto<?> writeReviewOfEvent(
-            @UserId Long userId,
-            @PathVariable Long eventId,
-            @RequestBody WriteReviewOfEvent writeReviewOfEvent
-    ){
-        eventService.writeReviewOfEvent(userId, eventId, writeReviewOfEvent);
-
-        return ResponseDto.ok(null);
-    }
-
     @GetMapping("{eventId}/top-review")
     public ResponseDto<?> showReviewsOfEvent(@PathVariable Long eventId){
         ReviewsOfEventDto reviewsOfEventDto = eventService.showReviewsOfEvent(eventId);
@@ -101,20 +72,6 @@ public class EventController {
         SellerOfEventDto sellerOfEventDto = eventService.showDetailInfoOfSeller(eventId);
 
         return ResponseDto.ok(sellerOfEventDto);
-    }
-    /**
-     * 사용자 티켓 구매하기
-     */
-    @PostMapping("{eventId}")
-    public ResponseDto<?> buyingTicket(
-            @UserId Long userId,
-            @PathVariable Long eventId,
-            @RequestBody BuyingTicketDto buyingTicketDto
-    ){
-        eventService.buyingTicket(userId, eventId, buyingTicketDto);
-
-        return ResponseDto.ok(null);
-
     }
 
     /**
